video_copy/requirements.txt
numpy==1.26.4
fastapi==0.111.0
Pillow==10.3.0
polars==0.20.31
pydantic==2.7.4
pydantic_settings==2.3.3
qdrant_client==1.9.1
Requests==2.32.3
timm==1.0.3
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
ultralytics==8.2.30
granian==1.4.3
